requirements.txt
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy

